pyproject.toml
[project]
name = "evolisa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.11.0",
    "cupy-cuda13x>=13.6.0",
    "ipython>=9.8.0",
    "isort>=7.0.0",
    "numpy>=2.3.5",
    "pillow>=12.0.0",
    "scikit-image>=0.25.2",
    "tqdm>=4.67.1",
]
